+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/input/MiniAppInputActivity.kt
@@ -12,7 +12,7 @@ import com.rakuten.tech.mobile.testapp.AppScreen.MINI_APP_INPUT_ACTIVITY
 import com.rakuten.tech.mobile.testapp.helper.isInvalidUuid
 import com.rakuten.tech.mobile.testapp.launchActivity
 import com.rakuten.tech.mobile.testapp.ui.display.MiniAppDisplayActivity
-import com.rakuten.tech.mobile.testapp.ui.display.PreloadMiniAppWindow
+import com.rakuten.tech.mobile.testapp.ui.display.preload.PreloadMiniAppWindow
 import com.rakuten.tech.mobile.testapp.ui.miniapplist.MiniAppListActivity
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import com.rakuten.tech.mobile.testapp.ui.settings.MenuBaseActivity
@@ -21,7 +21,7 @@ import com.rakuten.tech.mobile.testapp.ui.settings.SettingsMenuActivity
 class MiniAppInputActivity : MenuBaseActivity(), PreloadMiniAppWindow.PreloadMiniAppLaunchListener {
 
     private lateinit var binding: MiniAppInputActivityBinding
-    private val preloadMiniAppWindow by lazy { PreloadMiniAppWindow(this, this) }
+    private val preloadMiniAppWindow by lazy { PreloadMiniAppWindow(this,this) }
 
     sealed class InputDisplay(val input: String) {
         class AppId(input: String): InputDisplay(input)
